refactor(site): deduplicate shared utilities between messageParsing and streamState
Extract shared functions into new blockUtils.ts:
- mergeThinkingTitles: was duplicated verbatim in both files
- appendTextBlock: consolidates appendParsedTextBlock and
  appendStreamTextBlock using a joinText parameter to preserve
  the different text concatenation strategies (newline-joining
  for complete blocks, direct concatenation for streaming)
- asNonEmptyString: was duplicated in messageParsing.ts and
  chatHelpers.ts

Also adds a module-level doc comment to streamingJson.ts
describing its purpose and known limitations.
